cv-roaster/requirements.txt
flask==3.1.0
anthropic==0.49.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
